+++ docs/identifier_provisioning/CHANGELOG.md
@@ -2,8 +2,13 @@
 
 All notable changes to this gear are documented in this file.
 
-## 1.1.2
+## 1.1.4
+* Fixes a bug in retrieving module label from file suffix
 
+## 1.1.3
+* Adds `legacy` flag to enrollment record
+  
+## 1.1.2
 * Allows underscore in GUID
   
 ## 1.1.1